pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dartfx-fairproxy-api"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [{ name = "kulnor", email = "pascal.heus@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rdflib>=7.6.0",
    "requests-cache>=1.3.1",
    "uritools>=6.0.1",
    "uvicorn>=0.41.0",
]

[project.urls]
Documentation = "https://github.com/DataArtifex/dartfx-fairproxy-api#readme"
Issues = "https://github.com/DataArtifex/dartfx-fairproxy-api/issues"
Source = "https://github.com/DataArtifex/dartfx-fairproxy-api"


[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["src/dartfx"]

[tool.hatch.version]
path = "src/dartfx/fairproxy_api/__about__.py"

[dependency-groups]
dev = [
  "coverage[toml]>=6.5",
  "pytest>=7.0.0",
  "pytest-benchmark>=4.0.0",
  "mypy>=1.0.0",
  "python-dotenv>=1.0.0",
  "ruff>=0.3.0",
  "pre-commit>=3.0.0",
]
docs = [
  "sphinx>=7.0.0",
  "sphinx-rtd-theme>=1.2.0",
  "myst-parser>=1.0.0",
  "sphinx-copybutton>=0.5.0",
]

[tool.hatch.envs.default]
dependencies = ["dev"]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[[tool.hatch.envs.all.matrix]]
python = ["3.12", "3.13"]

[tool.hatch.envs.types]
dependencies = ["dev"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/fairproxy_api tests}"

[tool.hatch.envs.docs]
dependencies = ["docs"]
[tool.hatch.envs.docs.scripts]
build = "sphinx-build -b html docs/source docs/build/html"
serve = "python -m http.server -d docs/build/html"

[tool.hatch.metadata]
allow-direct-references = true

[tool.coverage.run]
source_pkgs = ["fairproxy_api", "tests"]
branch = true
parallel = true
omit = ["src/fairproxy_api/__about__.py"]

[tool.coverage.paths]
source = ["src/fairproxy_api", "*/dartfx/fairproxy-api/src/fairproxy_api"]
tests = ["tests", "*/dartfx/fairproxy-api/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.pytest.ini_options]
pythonpath = "src/dartfx"
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG", # flake8-unused-arguments
  "PT",  # flake8-pytest-style
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["dartfx"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
